+++ vck/src/commonMain/kotlin/at/asitplus/wallet/lib/jws/SelectiveDisclosureItemSerializer.kt
@@ -18,6 +18,7 @@ import kotlinx.serialization.json.jsonPrimitive
  * that is a JSON array with the values for salt, name, and the value itself,
  * which in turn can be anything, e.g. number, boolean or string.
  * We solve this challenge by serializing a list of [JsonPrimitive], see implementation.
+ * Note, that when disclosing array items, the claim name may be missing.
  */
 object SelectiveDisclosureItemSerializer : KSerializer<SelectiveDisclosureItem> {
 
@@ -28,23 +29,32 @@ object SelectiveDisclosureItemSerializer : KSerializer<SelectiveDisclosureItem> 
     override fun serialize(encoder: Encoder, value: SelectiveDisclosureItem) {
         encoder.encodeSerializableValue(
             listSerializer,
-            listOf(
+            listOfNotNull(
                 JsonPrimitive(value.salt.encodeToString(Base64UrlStrict)),
-                JsonPrimitive(value.claimName),
+                value.claimName?.let { JsonPrimitive(value.claimName) },
                 value.claimValue
             )
         )
     }
 
     override fun deserialize(decoder: Decoder): SelectiveDisclosureItem {
-        val items = decoder.decodeSerializableValue(listSerializer)
-        if (items.count() != 3) throw IllegalArgumentException()
-        val (firstElement, secondElement, thirdElement) = items
-        return SelectiveDisclosureItem(
-            salt = firstElement.jsonPrimitive.content.decodeToByteArray(Base64UrlStrict),
-            claimName = secondElement.jsonPrimitive.content,
-            claimValue = thirdElement
-        )
+        return with(decoder.decodeSerializableValue(listSerializer)) {
+            when (size) {
+                3 -> SelectiveDisclosureItem(
+                    salt = get(0).jsonPrimitive.content.decodeToByteArray(Base64UrlStrict),
+                    claimName = get(1).jsonPrimitive.content,
+                    claimValue = get(2)
+                )
+
+                2 -> SelectiveDisclosureItem(
+                    salt = get(0).jsonPrimitive.content.decodeToByteArray(Base64UrlStrict),
+                    claimName = null,
+                    claimValue = get(1)
+                )
+
+                else -> throw IllegalArgumentException("Neither 2 nor 3 elements")
+            }
+        }
     }
 
 }