+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Release 5.1.0:
    - Rename `disclosures` to `reconstructedJsonObject` in several result classes like `AuthnResponseResult.SuccessSdJwt`
    - Correctly implement confirmation claim in `VerifiableCredentialSdJwt`, migrating from `JsonWebKey` to `ConfirmationClaim`
    - Change type of `claimValue` in `SelectiveDisclosureItem` from `JsonPrimitive` to `JsonElement` to be able to process nested disclosures
+   - Implement deserialization of complex objects, including array claims
 
 Release 5.0.1:
  - Update JsonPath4K to 2.4.0